refactor(store): extract create selector to own file
